Ensure set index has tzone, if new value does not
Harvey Smith noticed that `index<-.xts` would return an object that
does not have a tzone attribute if the new value did not have one.

Set the tzone for the new index to be the same as the tzone for the
original index when the new index does not have a tzone attribute.

I considered adding a warning when this happens, but it probably
happens fairly often, since Ops.POSIXt drops the tzone attribute when
tzone = "". That would add a lot of noise.

Fixes #310.