Prepare for gcc 10 -fno-common default
All variable definitions need to be declared once, and defined once.

For whatever reason, this answer clicked for me:
https://stackoverflow.com/a/12511340/271616

  You should *declare* the variables, and be using `extern` to tell
  the compiler that the variables are defined somewhere else. Then in
  one source file you define the variables (i.e. the same as the
  declarations in the header file but without the `extern` keyword).

I initially marked all symbols and function definitions with `extern`.
But Martin Maechler pointed out that only symbols need to be declared
with `extern`. That reduced the number of changes to only affect the
global symbols and the imports from zoo.

Closes #316.