Fix first.default when n = -1
first.default() would always return the last two values of the (non-)
xts object, no matter how long the object was.

The unit tests did not catch this because they all ran on a vector of
three elements, so they didn't trigger the bug.

Also add a test for length-0 vectors at the beginning of the function,
like we do for the xts methods. Make the unit test vector 4 elements.

Fixes #325.