Another attempt to fix rownames
Setting .ROWNAMES caused some fts tests to fail because reclass()
called rownames<- on the xts object before calling re.fts(). And xts
objects can't have rownames, so the call was ignored and the reclassed
fts object didn't have rownames like the match.to object).

For reasons I don't understand, this call did not work.
    dimnames(xx)[[1]] <- NULL
It looks like it may be a bug in R...?

After those two changes, calls to Ops.xts() for '&' failed when it was
called on an xts object and a matrix with row names. This was because
.xts() was called on the result of the operation ('e') and 'e' had
rownames because the matrix did. Now we drop dimnames before
potentially calling .xts().

Fixes #298 (maybe)