Restore 'basic' ISO time format subsetting
The initial change for this fix only supports the extended format.
(hh:mm:ss.sss). We supported the basic format (hhmmss.sss) previously.
This restores the original functionality, making the colon optional.

I noticed how lastof() was being used to get the last timestamp for
the time string, and thought it would be good to find the first
timestamp in a similar manner. The getTimeComponents() function
extracts the hours, minutes, seconds, and sub-seconds into a list
that we can use with firstof() and lastof().

Note that firstof() returns the first timestamp in 1970, and lastof()
returns a timestamp at the end of 1970. The difference in years,
months, and days does not matter because we're only concerned with the
time of day components.

See #326. See #327.