Ensure endpoints() always returns last observation
endpoints() did not return the location of the last observation when
`k > 1` and `on` is "months" or "quarters". This was because the seq()
call did not include the last location.

Create a function to do the check, to keep things DRY.

Thanks to GitHub user Eluvias for the report.

Fixes #300.