Dispatch to window.zoo() for yearmon and yearqtr
Before 0.11-0, window.zoo() was dispatched when window() was called
on a xts object because there was no window.xts() method. window.zoo()
supports additional types of values for the `start` argument, and
possibly other features. So 0.11-0 had breaking changes I did not
anticipate.

Thanks to GitHub user annaymj for the report.

Fixes #312.