Fix "local var assigned but may not be used" NOTEs
From R CMD check:

  * checking R code for possible problems ... NOTE
  .xts: local variable ‘default.class’ assigned but...
  .xts: local variable ‘index.class’ assigned but...
  [<-.xts: local variable ‘.Class’ assigned but...
  addEventLines : <anonymous>: local variable ‘ta.x’ assigned but...
  addEventLines: local variable ‘ind’ assigned but...
  addLegend: local variable ‘ind’ assigned but...
  addPolygon : <anonymous>: local variable ‘ta.x’ assigned but...
  as.fts.xts: local variable ‘fts’ assigned but...
  as.timeSeries.xts: local variable ‘timeSeries’ assigned but...
  merge.xts: local variable ‘sfx’ assigned but...
  new.replot_xts: local variable ‘get_pad’ assigned but...
  new.replot_xts: local variable ‘move_frame’ assigned but...
  re.fts: local variable ‘fts’ assigned but...
  re.irts: local variable ‘irts’ assigned but...
  re.timeSeries: local variable ‘timeSeries’ assigned but...
  re.ts: local variable ‘dn’ assigned but...
  re.ts2 : na.replace: local variable ‘ncols’ assigned but...
  re.ts2: local variable ‘na.replace’ assigned but...
  rollcov.xts: local variable ‘n1’ assigned but...
  rollmax.xts: local variable ‘n1’ assigned but...
  rollmin.xts: local variable ‘n1’ assigned but...
  rollsum.xts: local variable ‘n1’ assigned but...
  tclass<-.default: local variable ‘tclass’ assigned but...
  tzone<-.default: local variable ‘tzone’ assigned but...

The two instances in .xts() needed to be added to the commented FIXME
code block. All the instances in the plot functions (add* and
new.replot_xts) were identified correctly and could be deleted without
issue.

The instances for local variables 'fts', 'irts', and 'timeSeries' are
due to the changes in cae4c9a50c84d30be8bf6b4afd2423cc2e9839f2,
"Reference package namespace in case it's not attached". The namespace
check intended to overwrite the constructor, but only if it was
attached. The commit changed the methods so they would work without the
package(s) attached, but only if the package(s) were available.

The instance in merge.xts() could be safely deleted. The re.ts2()
function was not used.

The instances in all the roll* functions were identified correctly and
could be deleted.

I'm not sure what to do about `[<-.xts`. The default replacement
methods for tclass and tzone look like bugs.