Allow adding new x-values within original range
Allow users to add a series to an existing plot when the new series
has x-values that do not exist in the original series. The only
restriction is that the new series cannot extend the x-axis beyond its
original start/end range.

The new behavior is consistent with chart_Series(). It looks like its
removal was a regression created when the 'observation.based = TRUE'
feature was implemented.

Fixes #360, see #216.