Allow panels to add any new x-axis values
Add a 'get_xcoords()' function and wait to set the x-coordinates until
the plot is drawn. That allows us to add series that have new index
values. It handles the cases when observation.based is either TRUE or
FALSE. You can pass an xts object as an argument in order to get the
x-coordinates that correspond to the observations in the object.

Add new functionality to '$update_frames()':

  1. The update_axis() function finds the unique index values across
     all the series that will be drawn, and then
  2. The update_lenv_data() function updates the series in each panel
     to make sure they all have at least one observation at every
     index value.

See #360.