Move disable warning note into the warning message
Sending the NOTE as a message() would print the message before the
warning. The message would be printed immediately, but the warning
wouldn't print until after the call finished. Making the warning
immediate caused it to print before the object was printed. That
would likely confuse users.

Add the NOTE to the warning the first time the warning is triggered.