Handle suffixes the same as merge() generic
merge.default() does not automatically add a "." between the column
names and the suffix like merge.zoo() does. So we have to choose
between being consistent with zoo or the generic.

This commit makes merge.xts() suffix handling consistent with the
generic because the generic is more flexible than automatically adding
a "." without a way to change the separator.

Fixes #371.