Set xlim and ylim before plot is rendered
Part of the change to allow users to add series with new x-axis values
was to delay setting the xlim values until the plot is rendered. This
is a regression from previous behavior.

Call update_frames() before getting xlim or ylim to ensure that their
values account for all series that are part of the plot at the time of
the call to get_xlim() or get_ylim().

Fixes #375.