Fix unintentional warnings during tests
Fix rbind() call in test-na.locf.R so it doesn't always convert 'xidx'
to numeric.

Add tests for expected warnings when calling .parseISO8601() with
garbage start and/or end times.

Fix partial argument match of:
    - 'tz' to 'tzone' in .xts() calls
    - 'index' to 'index.' in window.xts() calls

Suppress warnings when checking .indexCLASS, .indexFORMAT, and .indexTZ
arguments to the xts constructors. This uncovered a bug in the
constructors, which will be fixed in another commit.