Don't add index attributes to the xts object
The xts() constructor would add the .index* attributes to the xts
object via '...'. But those attributes should only be on the index,
not the xts object.

Also fix a bug where `missing("foo")` was used, but we should have
used `!hasArg("foo")` because missing is only for function arguments.
`missing("tzone")` was also a bug because it should have been
`missing(tzone)` (not quoted as a string).