Process index attributes before creating xts object
This moves all the index attribute processing to one place in the
constructors, which makes it easier to follow the logic. Add notes on
the precedence for the various index attributes possibilities.

Add tests to check that the precedence is correct. Also add tests to
make sure the timezone is set correctly for index classes that do not
have a timezone (i.e. equivalent to UTC).