Use time-based column if row name conversion fails
Users often have a data.frame with a date or timestamp column in
addition to all the data. We can look for that column if the 'rownames'
conversion fails, and spare the user from having to specify the date
column in 'order.by' and remove it when specifying 'x'.

Fixes #381.