Add tests for 'dimnames<-.xts'()
The GitHub issue this closes suggested that xts_set_dimnames() was
updating by reference when it shouldn't have. But it only updated by
reference if called via 'dimnames<-'(x, foo), which users shouldn't do.

Attempts to duplicate 'x' when it might have been shared with another
variable name caused 'x' not to be updated either.

The conclusion of this research is that 'dimnames<-.xts'() and
xts_set_dimnames() do not have bugs.

Also remove the check whether the 'value' passed to the dimnames setter
might be shared. The test after 'new_names[1] <- "e"' is to check
whether that line is needed or not.

Closes #292.