Fix merge() unit tests
The tests for "xm3" and "xm4" were comparing "xm2" again and again.

expect_identical(m1, m2) isn't correct because we want to make sure
they're equal to the zero_width_xts input, not each other (even though
that's the case).

Also switch the argument order, because the first is the current value
and the second is the target. We're targeting the zoo values.