Fix calculations for number of columns in output
This was wrong for a couple edge cases. Merging two or more zero-length
objects with column names returned an empty object. That's consistent
with zoo, but it's inconsistent with set algebra.

This change keeps the number of columns for zero-length objects, so we
have to process column names. (I also made sure that the result's
index has the attributes from the first argument, 'x').

See #222.