Make update_xaxis() a pure function
Updating 'x_axis' via <<- is potentially confusing because it looks
for 'x_axis' in the environment update_xaxis() is defined, not the
environment it was called from.

Also rename 'lenv' and 'lenv_data' to 'action_env' and 'action_data'
because those are more descriptive of what the object values represent.