Use the same code to set all y-axis labels
The y-axis labels were created differently for the main panel and
added panels. It's not clear why this is needed, and often caused
labels for added panels to be a different distance from the y-axis
than the main panel. This would be exaggerated if the plot window was
resized by the user.

Now the y-axis labels are created the same for the main panel and for
all panels added by addSeries().

Also move the y-axis labels half a character closer to the plot by
changing offset = 1 to 0.5.