Create panel object to hold header and series data
There was no distinction between header and series frames, but the
code assumed that headers and series were always in pairs. This
commit formalizes that assumption.

The render_panels() function combines set_window() and the loop over
the actions for each frame in plot.replot_xts().

Remove add_header() because the logic is now in new_panel(). Also
remove add_action() from the 'replot' object because it is now in the
'panel' object.

NOTES:

* The code still refers to the panels as 'frames'. They will be renamed
  in the next commit. That way this commit only has functional changes
  and is not cluttered with simple variable renames.

* Only the main plot method has been updated. addSeries() and friends
  still need to be migrated to this new structure.