Honor print.default() 'max' argument
print.xts() used 'max' to define how many rows were printed before and
after '...' when the output is truncated. This conflicts with the 'max'
argument in print.default() and breaks existing code.

Add a 'show.nrows' argument to provide this, instead of 'max'. And
ensure 'max' takes precedence over 'show.nrows' if it is specified.

Add documentation too.

Fixes #387.