Account for TZ when making names for split.xts()
as.yearmon.POSIXct() always sets tz = "GMT" when calling as.POSIXlt(),
regardless of the xts' index tzone. That can cause the as.yearmon()
results to be different days for GMT and the index's timezone.

Use format.POSIXct() for "months" because it checks for a tzone
attribute before converting to POSIXlt and calling format.POSIXlt().
The conversion to POSIXlt() is important because it checks and uses the
tzone attribute before considering the 'tz' argument. So it effectively
ignores the tz = "GMT" setting in as.yearmon(). This is also the reason
for calling as.POSIXlt() before calling as.yearqtr().

Fixes #392.