Add support for long vectors (> 2^32-1)
R added support for long vectors in version 3.0.0. The xts C code had
not been updated to take advantage of this feature. That meant xts
objects were limited to matrices with less than 2^32-1 elements.

Most of the changes are fall into these categories:

    1. Change index counters from 'int' to 'R_xlen_t'
    2. Use xlength() instead of LENGTH() or length() to get length
    3. Use xlengthgets() instead of lengthgets() to set length

There are a few non-functional changes to move type declaration closer
to where the variable is used.

See #239.