Register missing S3 methods and update signatures
With R-devel (83995-ish) `R CMD check` notes these S3 methods are not
registered. It also notes that the signatures for as.POSIXct.tis() and
str.replot_xts() do not match the respective generics.

It also thinks time.frequency() is a S3 method because time() is a
generic. The function isn't exported, so renaming won't break any
external code.

See #398.