Format each column individually before printing
The top/bottom rows could have a different number of decimal places
and there are often multiple variying spaces between columns. For
example:

                          close      volume          ma         bsi
2022-01-03 09:31:00     476.470  803961.000          NA   54191.000
2022-01-03 09:32:00     476.700  179476.000          NA   53444.791
2022-01-03 09:33:00     476.540  197919.000          NA  -16334.994
                ...
2023-03-16 14:52:00    394.6000  46728.0000    392.8636  28319.4691
2023-03-16 14:53:00    394.6500  64648.0000    392.8755  15137.6857
2023-03-16 14:54:00    394.6500  69900.0000    392.8873  -1167.9368

There are 4 spaces between the index and the 'close' column, 2 between
'close' and 'volume', 4 between 'volume' and 'ma', and 2 between 'ma'
and 'bsi'. There should be a consistent number of spaces between the
columns. Most other classes of objects print with 1 space between the
columns.

The top rows have 3 decimals and the bottom rows have 4. These should
also be the same.

See #321.