Update index if tclass set to non-TZ from TZ class
Updating the tclass from a class with a timezone to one without a
timezone did not actually update the index. This meant that converting
a POSIXct index to a Date index wouldn't actually change the index to
a POSIXct value at midnight. Thanks to Daniel Palomar for the report.

Reverse dependency checks passed after making this change.

Fixes #311.