Fix multi.panel functionality
This wasn't updated to use the new panel functionality. cs$add_action()
and cs$add_header() no longer exist.

Moved everything into the for-loop because we can add the first series
to the main panel the same way we add the remaining panel series. This
could probably be refactored and simplified, but I'll leave that for a
later commit.