Track last action/render panel separately
We need to know which panel was the most recent one:
  1) that was rendered (already implemented), and
  2) an action was added to (not implemented).

The `Env$panel` value tracked the last panel rendered. We need this to
make sure get_ylim() retrieves the ylim for the correct panel.

We also need to track the panel that the last action was added to. We
need this so calls respect the values for 'on' when it is not NA. For
example, the last active panel should be '1' in the code below:

    plot(x)
    addSeries(y, on = NA) # new panel, active = 2
    addSeries(z, on =  1) # panel 1, active = 1

Add an 'id' to each panel so it knows where it is on the plot. Then we
can set the last_active_panel_id any time add_action() is called on a
panel.

Also fix panel updates when 'on' is a vector. This was a silly mistake.
It would error if 'on' was a vector because get_panel() only works for
a scalar value.

    for(i in seq_along(on)) {
        this_panel <- plot_object$get_panel(on)