Only draw x-axis once when multi.panel = TRUE
We were re-drawing the x-axis every time a panel was rendered. This
caused the axis annotations look like a 'bold' font. Adding the x-axis
was only being used for its side-effect of calling get_ylim().
Otherwise the 'ylim' of the panel could be c(Inf, -Inf) due to
re-calculating every panel's 'ylim' during rendering.

Instead, add a get_ylim() expression to the panel when new_panel() is
called. Require the 'envir' argument to new_panel() so the call to
get_ylim() is evaluated in the correct environment.