Refactor logic for when actions update ylim
Sometimes we don't want an action to update a panel's ylim based on
the data in the action.

Always add an attribute to the action to indicate whether or not it
should be able to update ylim. 'update_ylim = TRUE' by default. Set
'update_ylim = FALSE' instead of 'no.update = TRUE' when we don't want
the action to change the panel ylim.

Also remove the 'no_update' argument to new_panel() since it was just
being used as a hard-coded value inside the function, and never set by
any call to new_panel().