Fix ylim update when multi.panel = TRUE
This threw an error because only the first panel's ylim was updated by
the get_ylim() call in the xaxis expression. All other panels' ylim was
still c(Inf, -Inf) which caused an error when drawing them. The call to
get_ylim() in the new_panel() constructor did not fix this issue.

Add a 'ylim_render' field to the panel object so we do not overwrite
the original 'ylim' values. That way the original values will always
be available when we need to determine the 'ylim_render' value based
on the original value and the min/max values of the data.