Fix multi-panel header spacing
This makes sure 7% of the plot window is reserved for the main plot
header, and 5% is reserved for each panel header. This makes sure the
headers always have the same amount of space on the plot, regardless of
how many series there are. Otherwise the header text could be covered
by the series plot.