Simplify panel logic with and without multi-panels
Now that the main plot header and x-axis are in the global plot
environment, we do not need to create the main panel before we create
the remaining panels for a multi-panel plot.

Create a new panel the same way for every multi-panel in a single
plot. Move the main panel creation to the multi-panel 'else' block
instead of creating it before the multi-panel if statement.

This also allows us to use the same header logic for all panels,
regardless of whether or not they're part of a multi-panel plot.
Rename the 'header_type' argument to 'header' and always specify a
header when creating a new panel.