Remove 'use_get_ylim' and use 'is_ylim_fixed'
Multi-panel plots were the only time 'use_get_ylim' was TRUE. And the
only time it had any effect was when 'is_ylim_fixed' was also FALSE,
which was only true when 'yaxis.same' was TRUE.

This was one of two times we had to re-calculate the 'ylim' based on
every panel. The other time was when 'ylim' is NULL and 'multi.panel'
was FALSE. And 'is_ylim_fixed' is also FALSE in that case, so we know
to re-calculate 'ylim' whenever 'is_ylim_fixed' is FALSE.

See #103.