Remove 'update_ylim' attribute from actions
The 'ylim' belongs to the panel, not each individual action. The panel
constructor's 'use_fixed_ylim' argument determines whether or not the
panel 'ylim' can be modified by an action.

This changes how 'ylim' is determined when a new panel is added to a
plot. Previously, the 'ylim' for the new panel was calculated as the
range of the data provided when the panel was created, and the 'ylim'
for the new panel was always fixed. Adding more series to an existing
panel would not update the panel 'ylim' value, so added series values
outside the initial 'ylim' would not appear on the panel.