Allow negative values for 'on'
get_panel() threw an error when 'on' is negative because you cannot
subset a list with a negative value. When 'on' is negative, the added
action should be rendered before all existing series actions.

Add function add_panel_action() to handle negative values for 'on'.
Update the panel's add_action() function to append the new action to
the correct place in the 'actions' list. The header action is always
the panel's first action, and the grid lines should be rendered before
all series actions.

Aside: the "background" value for 'where' will handle cases where the
action should be rendered before the grid lines. For example, when
adding shading for a time range (e.g. recession shading).

Fixes #409.