Position plot header with 'adj'
The use of 'pos' and 'offset' can cause the top of the header to be
clipped off and not visible if the plot space is not large enough
vertically. This can happen when the device is resized, or if multiple
plots are rendered on one device using layout().

Thanks to Dirk Eddelbuettel for the report.

Fixes #412.