Ensure 'i' is integer and warn if it isn't
Convert 'i' to integer in the R code to avoid issues when it's passed
to the C code. I couldn't get this to crash like I could with the code
for 'j', but it caused incorrect results when values in 'i' were set
to zero during conversion to integer.

See #413.