Move merge C code to macro
This allows us to call the main merge algorithm inside a switch
statement, instead of calling the switch statements inside the main
loop. Tests showed a roughly 10% speed improvement.

This macro does not--and should not--rely on any variables that aren't
defined as macro arguments. This isn't enforceable by the compiler, so
all the variables used have an underscore prefix and suffix to avoid
referencing a value outside of the macro's intended scope.